Separate Static decoder tests
